Moved DSL to specific package
