+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/core/command/Commands.kt
@@ -4,8 +4,8 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import dev.syoritohatsuki.duckyupdaterrework.core.command.execute.check
 import dev.syoritohatsuki.duckyupdaterrework.core.command.execute.update
 import dev.syoritohatsuki.duckyupdaterrework.core.command.execute.updateAll
-import dev.syoritohatsuki.duckyupdaterrework.util.argument
-import dev.syoritohatsuki.duckyupdaterrework.util.literal
+import dev.syoritohatsuki.duckyupdaterrework.dsl.argument
+import dev.syoritohatsuki.duckyupdaterrework.dsl.literal
 import net.minecraft.command.CommandSource
 
 fun LiteralArgumentBuilder<out CommandSource>.commands() {
