+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/server/DuckyUpdaterReWorkServer.kt
@@ -2,8 +2,8 @@ package dev.syoritohatsuki.duckyupdaterrework.server
 
 import dev.syoritohatsuki.duckyupdaterrework.DuckyUpdaterReWork.logger
 import dev.syoritohatsuki.duckyupdaterrework.core.command.commands
-import dev.syoritohatsuki.duckyupdaterrework.util.register
-import dev.syoritohatsuki.duckyupdaterrework.util.rootLiteral
+import dev.syoritohatsuki.duckyupdaterrework.dsl.register
+import dev.syoritohatsuki.duckyupdaterrework.dsl.rootLiteral
 import net.fabricmc.api.DedicatedServerModInitializer
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback
